fix(pg-cfpages-basic): fix `pg` compatibility with CF Pages (#5960)
`pg` code fails to bundle for Cloudflare Pages due to the dependencies
on built-in Node.js modules that are imported in an outdated way:

```
  ✘ [ERROR] Could not resolve "events"

      ../node_modules/.pnpm/pg-cloudflare@1.1.1/node_modules/pg-cloudflare/dist/index.js:1:29:
        1 │ import { EventEmitter } from 'events';
```

Previously this worked when building with `--compatibility-flags
"nodejs_compat"` but now it doesn't anymore, because these imports now
require the explicit `node:` prefix (e.g. `import { EventEmitter } from
'node:events'`).

This commit swaps `pg` with `@prisma/pg-worker` which is a drop-in
replacement for `pg` that works with Cloudflare Workers and Pages. It
still requires the `nodejs_compat` flag because it uses some builtin
Node.js modules but they all are correctly imported using the `node:`
protocol.

Fixes:
https://linear.app/prisma-company/issue/ORM-846/fix-driver-adapters-wasm-pg-cfpages-basic-wasm-ubuntu-2204